Move world preview image from overlay
