+++ src/ui/views/session/world/World.tsx
@@ -1,0 +1,11 @@
+import { WorldThumbnail } from "./WorldThumbnail";
+import { WorldView } from "./WorldView";
+
+export default function World() {
+  return (
+    <>
+      <WorldView />
+      <WorldThumbnail />
+    </>
+  );
+}
