+++ src/ui/views/session/world/WorldView.tsx
@@ -112,7 +112,7 @@ const MuteButton = forwardRef<HTMLButtonElement, { activeCall?: GroupCall; showT
   }
 );
 
-export default function WorldView() {
+export function WorldView() {
   const { canvasRef, world, onExitWorld, onWorldTransfer, activeCall } = useOutletContext<SessionOutletContext>();
   const isEnteredWorld = useStore((state) => state.world.isEnteredWorld);
   const { isOpen: isChatOpen, openWorldChat, closeWorldChat } = useStore((state) => state.worldChat);
